# conf/openapi.yaml
openapi: 3.1.0
info:
  title: seblm-meals API
  version: 1.0.0
tags:
  - name: Meal
servers:
  - url: http://localhost:9000/api
paths:
  /meal/{id}:
    get:
      tags:
        - Meal
      description: Get meal by id
      parameters:
        - name: id
          description: Unique identifier of the meal
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealStatistics'
          description: Meal statistics
  /meals/statistics:
    get:
      tags:
        - Meal
      description: Get all meals
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealStatisticsResponse'
          description: All meals
  /meals/{year}/{week}:
    get:
      tags:
        - Meal
      description: Get all meals for a week
      operationId: mealsApi
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: week
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: WeekMeals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeekMeals'
  /meals/{year}-{month}-{day}:
    get:
      tags:
        - Meal
      description: Get meals centered around a day
      operationId: mealsAround
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
        - name: day
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: WeekMealsCenteredAroundADay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeekMealsCenteredAroundADay'
  /suggest:
    get:
      tags:
        - Meal
      description: Suggest some meals according to a reference date and an optional search
      operationId: suggest
      parameters:
        - name: reference
          in: query
          required: true
          description: Format is either 'YYYY-MM-DD 12:00' or 'YYYY-MM-DD 20:00'
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuggestResponse'
          description: Meal suggestions
  /link-or-insert:
    post:
      tags:
        - Meal
      description: Create a meal
      operationId: linkOrInsertApi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkOrInsertData'
  /unlink:
    delete:
      tags:
        - Meal
      description: Unlink a meal
      operationId: unlinkApi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlinkMeal'
components:
  schemas:
    MealStatistics:
      type: object
      required:
        - count
      properties:
        count:
          type: number
        first:
          type: string
        last:
          type: string
        meal:
          $ref: '#/components/schemas/Meal'
    MealStatisticsResponse:
      type: array
      items:
        $ref: '#/components/schemas/MealStatistics'
    WeekMeals:
      required:
        - titles
        - previous
        - now
        - next
        - monday
        - tuesday
        - wednesday
        - thursday
        - friday
        - saturday
        - sunday
      type: object
      properties:
        titles:
          $ref: '#/components/schemas/Titles'
        previous:
          $ref: '#/components/schemas/WeekReference'
        now:
          $ref: '#/components/schemas/WeekReference'
        next:
          $ref: '#/components/schemas/WeekReference'
        monday:
          $ref: '#/components/schemas/WeekDay'
        tuesday:
          $ref: '#/components/schemas/WeekDay'
        wednesday:
          $ref: '#/components/schemas/WeekDay'
        thursday:
          $ref: '#/components/schemas/WeekDay'
        friday:
          $ref: '#/components/schemas/WeekDay'
        saturday:
          $ref: '#/components/schemas/WeekDay'
        sunday:
          $ref: '#/components/schemas/WeekDay'
    Titles:
      required:
        - short
        - long
      type: object
      properties:
        short:
          type: string
        long:
          type: string
    WeekReference:
      required:
        - year
        - week
        - isActive
      type: object
      properties:
        year:
          type: integer
        week:
          type: integer
        isActive:
          type: boolean
    WeekDay:
      required:
        - reference
      type: object
      properties:
        reference:
          type: string
          format: date
        lunch:
          $ref: '#/components/schemas/MealEntry'
        dinner:
          $ref: '#/components/schemas/MealEntry'
    WeekMealsCenteredAroundADay:
      required:
        - days
      type: object
      properties:
        days:
          type: array
        items:
          $ref: '#/components/schemas/WeekDay'
    Meal:
      required:
        - description
        - id
      type: object
      properties:
        description:
          type: string
        id:
          type: string
          format: uuid
        url:
          type: string
    MealEntry:
      required:
        - meal
        - time
      type: object
      properties:
        meal:
          $ref: '#/components/schemas/Meal'
        time:
          type: string
    SuggestResponse:
      type: object
      properties:
        fiftyTwoWeeksAgo:
          $ref: '#/components/schemas/MealSuggest'
        fourWeeksAgo:
          $ref: '#/components/schemas/MealSuggest'
        mostRecents:
          type: array
          items:
            $ref: '#/components/schemas/MealSuggest'
    MealSuggest:
      required:
        - count
        - description
        - descriptionLabel
        - lastused
      type: object
      properties:
        count:
          type: integer
        description:
          type: string
        descriptionLabel:
          type: string
        lastused:
          type: integer
    LinkOrInsertData:
      required:
        - mealDescription
        - mealTime
      type: object
      properties:
        mealDescription:
          type: string
        mealTime:
          type: string
          format: date-time
    UnlinkMeal:
      required:
        - mealTime
      type: object
      properties:
        mealTime:
          type: string
          format: date-time